src/main/java/com/epam/hotel/command/implementation/LogOutCommand.java
package com.epam.hotel.command.implementation;

import com.epam.hotel.command.Command;
import com.epam.hotel.command.CommandResult;
import com.epam.hotel.exception.ServicesException;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.http.HttpSession;

public class LogOutCommand implements Command {
    private static final String MAIN_PAGE = "WEB-INF/view/main.jsp";

    @Override
    public CommandResult execute(HttpServletRequest request, HttpServletResponse response) throws ServicesException {
        HttpSession session = request.getSession();
        session.invalidate();
        return CommandResult.forward(MAIN_PAGE);
    }
}
